Added Inspector blocked statistics.

+++ sim/src/main/java/Inspector.java
@@ -52,17 +52,21 @@ public class Inspector {
         return depart;
     }
 
-    public String get(double clock) {
-        // Get the component in service
-        String component = this.inService.remove(0);
+    public void get(double clock) {
         
-        // Update clock and statistics
         this.clock = clock;
-        this.numInService = 0;
-        this.totalBusy += (this.clock - this.lastEventTime);
-        this.lastEventTime = this.clock;
         
-        return component;
+        // If this Inspector is blocked, just update total blocked
+        if (this.isBlocked) {
+            this.totalBlocked += (this.clock - this.lastEventTime);
+        } else {
+            // Remove the component from service
+            this.inService.remove(0);
+            // Update statistics
+            this.numInService = 0;
+            this.totalBusy += (this.clock - this.lastEventTime);
+        }
+        this.lastEventTime = this.clock;
     }
 
     public Event scheduleDeparture(String component) {
@@ -100,6 +104,19 @@ public class Inspector {
     }
 
     public void qReportGeneration(double clock) {
-        throw new NotImplementedException();
+        this.clock = clock;
+        
+        // Update statistics
+        if (this.isBlocked)
+            this.totalBlocked += (this.clock - this.lastEventTime);
+        else
+            this.totalBusy += (this.clock - this.lastEventTime);
+        
+        double pBlocked = this.totalBlocked / this.clock;
+        
+        System.out.printf("*** INSPECTOR %s REPORT ***\n", this.id);
+        System.out.printf("Total time busy = %.1f\n", this.totalBusy);
+        System.out.printf("Total time blocked = %.1f\n", this.totalBlocked);
+        System.out.printf("Probability of being blocked = %.1f\n\n", pBlocked);
     }
 }
